batch file updates

+++ Windows/Code/pharaoh-windows.bat
@@ -1,12 +1,12 @@
 @ECHO OFF
 REM BFCPEOPTIONSTART
 REM Advanced BAT to EXE Converter www.BatToExeConverter.com
-REM BFCPEEXE=C:\Users\vagrant\native-installers\Windows\pharaohtools2.exe
-REM BFCPEICON=
+REM BFCPEEXE=\\VBOXSVR\Windows_Native_Installer_Files\pharaohtools.exe
+REM BFCPEICON=\\VBOXSVR\Windows_Native_Installer_Files\Code\logo-pharaoh-high.ico
 REM BFCPEICONINDEX=1
-REM BFCPEEMBEDDISPLAY=0
+REM BFCPEEMBEDDISPLAY=1
 REM BFCPEEMBEDDELETE=1
-REM BFCPEADMINEXE=0
+REM BFCPEADMINEXE=1
 REM BFCPEINVISEXE=0
 REM BFCPEVERINCLUDE=1
 REM BFCPEVERVERSION=1.0.0.0
@@ -14,38 +14,76 @@ REM BFCPEVERPRODUCT=Pharaoh Tools
 REM BFCPEVERDESC=PHP Automation
 REM BFCPEVERCOMPANY=Pharaoh Tools
 REM BFCPEVERCOPYRIGHT=
-REM BFCPEEMBED=C:\Users\vagrant\native-installers\Windows\Code\wget.exe
-REM BFCPEEMBED=C:\Users\vagrant\native-installers\Windows\Code\unzip.exe
-REM BFCPEEMBED=C:\Users\vagrant\native-installers\Windows\Code\vc_redist.x86.exe
+REM BFCPEEMBED=\\VBOXSVR\Windows_Native_Installer_Files\Code\wget.exe
+REM BFCPEEMBED=\\VBOXSVR\Windows_Native_Installer_Files\Code\vc_redist.x86.exe
+REM BFCPEEMBED=\\VBOXSVR\Windows_Native_Installer_Files\Code\unzip.exe
+REM BFCPEEMBED=\\VBOXSVR\Windows_Native_Installer_Files\SourceFiles\Windows6.1-KB2999226-x86.msu
+REM BFCPEEMBED=\\VBOXSVR\Windows_Native_Installer_Files\SourceFiles\Windows6.0-KB2999226-x86.msu
+REM BFCPEEMBED=\\VBOXSVR\Windows_Native_Installer_Files\SourceFiles\Windows8.1-KB2999226-x86.msu
+REM BFCPEEMBED=\\VBOXSVR\Windows_Native_Installer_Files\SourceFiles\Windows8-RT-KB2999226-x86.msu
 REM BFCPEOPTIONEND
 @ECHO ON
 @echo off
+REM Start Nicely
 echo "Our Pharaoh Tools Windows Executable"
-timeout 5
-del /S /Q %TEMP%\php.zip
+REM Tell Them we're about to start
+timeout 10
+REM Copy wget from installer to temp dir so we can wget PHP
 echo "Copying wget from installer to %TEMP%\wget.exe"
 copy %MYFILES%\wget.exe %TEMP%\wget.exe
+REM Go to the Temp dir
 echo "Changing directory to %TEMP%"
 cd %TEMP%
+REM Do the PHP Download 
 echo "Downloading PHP with wget"
 wget.exe http://windows.php.net/downloads/releases/archives/php-7.0.2-Win32-VC14-x86.zip
+REM Deleting %SystemDrive%\php.zip if it exists
 echo "Deleting %SystemDrive%\php.zip if it exists"
 del /S /Q %SystemDrive%\php.zip
+REM Overwrite %SystemDrive%\php.zip with fresh copy
 echo "Copying downloaded php from %TEMP%\php-7.0.2-Win32-VC14-x86.zip to %SystemDrive%\php.zip"
 copy %TEMP%\php-7.0.2-Win32-VC14-x86.zip %SystemDrive%\php.zip
+REM Delete the php dir if its there
 echo "Deleting %SystemDrive%\php if it exists"
 rmdir /S /Q %SystemDrive%\php
+REM Create new empty php directory
 echo "Making directory %SystemDrive%\php"
-mkdir %SystemDrive%\php
+if not exist "%SystemDrive%\php" mkdir %SystemDrive%\php
+REM CD to php dir
 echo "Changing directory to %SystemDrive%\php"
 cd %SystemDrive%\php
+REM Check our current directory, get the unzip exe
 echo "Current directory is %cd%"
-echo "Unzipping to %SystemDrive%\php.zip"
+echo "Copying unzip from installer to %SystemDrive%\unzip.exe"
+copy %MYFILES%\unzip.exe %SystemDrive%\unzip.exe
+REM Unzip php into the php dir
+echo "Unzipping %SystemDrive%\php.zip"
 %SystemDrive%\unzip.exe %SystemDrive%\php.zip
+REM Get rid of the original php zip
 del /S /Q %SystemDrive%\php.zip
+REM Install the Visual Studio redistributable that php 7 needs
 echo "Installing Visual Studio Redistributable"
-start /wait %MYFILES%\vc_redist.x86.exe /install /quiet /norestart
-REM install chocolatey
+start /wait %MYFILES%\vc_redist.x86.exe /install /passive /norestart
+REM Get the right version to install the right MSU
+setlocal
+for /f "tokens=4-5 delims=. " %%i in ('ver') do set VERSION=%%i.%%j
+if "%version%" == "6.3" (
+    echo "Found Windows Version: Windows 8.1"
+    wusa.exe %MYFILES%\Windows8.1-KB2999226-x86.msu
+) else if "%version%" == "6.2" (
+    echo "Found Windows Version: Windows 8"
+    wusa.exe %MYFILES%\Windows8-RT-KB2999226-x86.msu
+) else if "%version%" == "6.1" (
+    echo "Found Windows Version: Windows 7"
+    wusa.exe %MYFILES%\Windows6.1-KB2999226-x86.msu
+) else if "%version%" == "6.0" (
+    echo "Found Windows Version: Windows Vista"
+    wusa.exe %MYFILES%\Windows6.0-KB2999226-x86.msu
+) else (
+    echo "Unable to calculate Windows Version to ensure Updates for "
+)
+endlocal
+REM install Chocolatey
 echo "Installing the Chocolatey Package Manager for Windows"
 start /wait cmd /C @powershell -NoProfile -ExecutionPolicy Bypass -Command "iex ((new-object net.webclient).DownloadString('https://chocolatey.org/install.ps1'))" && SET PATH=%PATH%;%ALLUSERSPROFILE%\chocolatey\bin
 start /wait cmd /C choco install nuget.commandline -y
@@ -56,16 +94,19 @@ REM Download Pharaoh configure
 echo "Downloading the latest version of Pharaoh Configure"
 echo %TEMP%\ptconfigure-install
 timeout 5
-rmdir /S /Q %TEMP%\ptconfigure-install
+REM Get rid of the ptconfigure install files if they do exist
+if exist "%TEMP%\ptconfigure-install" rmdir /S /Q %TEMP%\ptconfigure-install
+REM clone 
 start /wait cmd /C git clone https://github.com/PharaohTools/ptconfigure %TEMP%\ptconfigure-install
 timeout 5
 REM Install Pharaoh configure
 echo "Installing the latest version of Pharaoh Configure"
-start /wait cmd /C %SystemDrive%\php\php.exe%TEMP%\ptconfigure-install\install-silent
+start /wait cmd /C "%SystemDrive%\php\php.exe %TEMP%\ptconfigure-install\install-silent && timeout 30"
 REM Install Pharaoh Tools
 echo "Installing all available Pharaoh Tools"
 REM php Ptconfigure pharaohtools install -yg
 start /wait cmd /C %SystemDrive%\php\php.exe ptconfigure pharaohtools install -yg
 REM Thank You for installing message
 echo "Thank you for Installing the latest version of Pharaoh Tools"
+REM Ending
 timeout 15
